refactor: EquipmentItem 단위 컨트롤러/서비스 책임 분리
- 책임 분리 하는 과정에서 uri 변경 : /equipment-items/....
