+++ src/main/java/com/equip/equiprental/equipment/service/EquipmentService.java
@@ -3,7 +3,6 @@ package com.equip.equiprental.equipment.service;
 import com.equip.equiprental.common.dto.PageResponseDto;
 import com.equip.equiprental.common.dto.SearchParamDto;
 import com.equip.equiprental.equipment.dto.*;
-import com.equip.equiprental.member.domain.Member;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
@@ -15,9 +14,6 @@ public interface EquipmentService {
     EquipmentItemListDto getEquipmentItem(Long equipmentId, SearchParamDto paramDto);
 
     void increaseStock(Long equipmentId, IncreaseStockRequestDto dto);
-    void updateItemStatus(UpdateItemStatusDto dto, Member changer);
-
-    PageResponseDto<EquipmentItemHistoryDto> getItemHistory(Long equipmentItemId, SearchParamDto paramDto);
 
     void updateEquipmentImage(Long equipmentId, List<MultipartFile> files);
 }
